+++ esb_prx_blefallback/CMakeLists.txt
@@ -8,9 +8,9 @@ cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(esb_prx_blefallback)
 
-zephyr_include_directories(.) # ble
+zephyr_include_directories(.) # ble, io
 
-FILE(GLOB app_sources src/*.c src/ble/*.c)
+FILE(GLOB app_sources src/*.c src/ble/*.c src/io/*.c)
 # NORDIC SDK APP START
 target_sources(app PRIVATE ${app_sources})
 # NORDIC SDK APP END
